PRSD-901: Move iterator to a separate file
